+++ .gitignore
@@ -4,5 +4,6 @@ __pycache__
 .ipynb_checkpoints
 venv
 venv2
+venv3
 src/keys.py
 .idea
