+++ docs/pyproject.toml
@@ -3,5 +3,5 @@ requires = ["requests >= 2.27.1", "DateTime >= 4.3"]
 
 [project]
 name = "jockmkt-sdk"
-authors = [{name = "Alexander Friedman/nysugfx", email = "nysu.gfx@gmail.com"}]
+authors = [{name = "Alex Friedman/nysugfx", email = "nysu.gfx@gmail.com"}]
 dynamic = ["version", "description"]
