+++ setup.cfg
@@ -1,2 +1,30 @@
 [metadata]
+name = jockmkt-sdk
+version = 0.1.2
+author = Alex Friedman
+author_email = nysu.gfx@gmail.com
+description = A basic package allowing the user to interact with Jock MKTs API
+long_description = file: README.rst
+long_description_content_type = text/x-rst
+url = https://github.com/nysugfx/jockmkt-sdk
+project_urls =
+    BugTracker = https://github.com/nysugfx/jockmkt-sdk/issues
+classifiers =
+    Intended Audience :: Developers
+    Topic :: Software Development :: Libraries :: Python Modules
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3.10
+    Operating System :: OS Independent
 license_files = LICENSE.txt
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.10
+install_requires =
+    requests >= 2.27.1
+    DateTime >= 4.3
+
+[options.packages.find]
+where = src
