+++ src/jockmkt_sdk/exception.py
@@ -3,12 +3,14 @@ from datetime import datetime, timedelta
 
 class JockAPIException(Exception):
     """
-    code -- type -- explanation 400 -- bad_request -- required parameter not included or typo in required parameter
-    401 -- not_authorized -- token is not valid; expired or keys not functioning 402 -- request_failed -- event
-    status error, user has not joined event, insufficient funds 404 -- not_found -- incorrect api endpoint 429 --
-    rate_limit -- max 10 orders (post, delete) per minute, max 250 other requests per minute. This limit resets at
-    the beginning of every new clock minute (e.g 12:00:00, 12:01:00) 50x -- internal_error -- request failed due to
-    platform error
+    code -- type -- explanation
+    400 -- bad_request -- required parameter not included or typo in required parameter
+    401 -- not_authorized -- token is not valid; expired or keys not functioning
+    402 -- request_failed -- event status error, user has not joined event, insufficient funds
+    404 -- not_found -- incorrect api endpoint
+    429 -- rate_limit -- max 10 orders (post, delete) per minute, max 250 other requests per minute. This limit resets
+    at the beginning of every new clock minute (e.g 12:00:00, 12:01:00)
+    50x -- internal_error -- request failed due to platform error
     """
 
     def __init__(self, response):
@@ -18,12 +20,11 @@ class JockAPIException(Exception):
             'bad_request': "try fixing your parameters, or check if you're missing one!",
             'not_authorized': 'Double check your secret keys, or that your auth token is valid',
             'event_status': 'Please wait for the market to open at ...',
-            'invalid entry': '',  # TODO: create entry to the event
+            'invalid entry': '',
             'not_found': '',
             'rate_limit': 'You have placed too many orders or requests since {}. Please wait until {}.'.format(
                 now, rate_limit_reset),
             'request_failed': 'You have already entered the event or deleted your order'
-            # TODO: store the request so that it can be completed at the start of the following minute
         }
         self.code = ""
         self.message = 'unknown error'
@@ -44,3 +45,12 @@ class JockAPIException(Exception):
 
     def __str__(self):
         return 'JockAPIException {}: {} \n{}'.format(self.code, self.message, self.helper)
+
+# class JockInputException(Exception):
+#     _LEAGUES = []
+#     _LEN_API_KEY = 23
+#     _LEN_SECRET = 32
+#     _LEN_IDS = 37
+#
+#     def __init__(self, **kwargs):
+#         self.input_league = kwargs.get('league')
